+++ src/components/Input.module.css
@@ -1,0 +1,25 @@
+.box > * {
+    display: block;
+}
+
+.box label {
+    margin-bottom: 16px;
+    font-weight: 700;
+    font-size: 18px;
+    line-height: 26px;
+}
+
+.box input {
+    width: 100%;
+    padding: 15px 24px ;
+    border: none;
+    border-radius: 12px;
+    background-color: var(--gray100);
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 26px;
+}
+
+.box input::placeholder {
+    color: var(--gray400);
+}
